Refactor Hire me Modal Relates #101
