--- client/src/Pages/Orders/style.css
@@ -1,118 +1,0 @@
-.Orders {
-  position: relative;
-  min-height: 400px;
-  margin: 0 !important;
-  /* width: 700px; */
-}
-.Orders-card__details {
-  text-align: left;
-  padding-top: 6rem;
-  color: #485C73;
-}
-.Orders .bg-image {
-  background-position: center;
-  background-size: cover;
-  position: absolute;
-  height: 160px;
-  width: 100%;
-}
-
-.image-container {
-  position: relative;
-  min-height: 250px;
-  width: 180px;
-  border-radius: 3px;
-}
-
-.image-container span {
-  display: inline-block;
-  vertical-align: middle;
-  position: absolute;
-  width: 100%;
-  background-color: #f0b240af;
-  color: #f5f5f5;
-  bottom: 0;
-  height: 40px;
-  font-size: 22px;
-  line-height: 40px;
-  border-bottom-right-radius: 3px;
-  border-bottom-left-radius: 3px;
-  left: calc(50% - 50%);
-  text-align: center;
-}
-
-
-.Orders .Orders-card {
-  margin: 4.5rem 3rem;
-}
-
-.Orders-details {
-  width: 100%;
-}
-
-.label-text {
-  font-size: 12px;
-}
-
-.label-text__rating {
-  padding-right: 5px;
-}
-
-.rating {
-  font-size: 16px;
-}
-
-.label-text__emp {
-  color: #252525;
-  font-weight: bold;
-}
-
-.ant-typography {
-  font-size: 16px;
-}
-
-/* .Orders-detail-input {
-  margin-top: 2rem;
-} */
-
-/* .hireme-btn {
-  width: 100%;
-  font-size: 24px;
-  font-weight: normal;
-  border-radius: 3px;
-  margin-top: 1rem;
-} */
-
-/* .signin-link {
-  display: inline-block;
-  margin-top: 0.5rem;
-  color: white;
-  border: 1px solid #FDCB6E;
-  background-color: #FDCB6E;
-  border-radius: 3px;
-  padding-left: 0;
-  width: 200px;
-  height: 45px;
-  line-height: 45px;
-  text-align: center;
-  transition: all 0.25s;
-}
-
-.signin-link:hover {
-  color: white;
-  background-color: transparent;
-  transition: all 0.25s;
-} */
-
-/* .loggedin-message {
-  text-align: center;
-} */
-
-@media only screen and (max-width: 768px) {
-  .image-container {
-    margin: 0 auto;
-  }
-  .Orders-card__details {
-    padding-top: 2rem;
-  }
-}
